% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_area.R
\name{total_area}
\alias{total_area}
\title{Compute total area of a landscape}
\usage{
total_area(landscape)
}
\arguments{
\item{landscape}{SpatRaster, the landscape raster (terra SpatRaster object)}
}
\value{
Total area in hectares (ha), omitting NA values.
}
\description{
Compute total area of a landscape
}
\examples{
library(terra)
landscape <- terra::rast(matrix(1, 10, 10)) # example raster
total_area(landscape)
}
